// src/app/(routes)/(refacciones)/gestion_inventario_refacciones/components/TablaRefacciones/TablaRefacciones.types.ts
export type Refaccion = {
  codigo: number
  descripcion: string
  noParte: string
  existenciaFisica: number
  existenciaSistema: number
  diferencias: number
  proveedores: string
  cantidadEntrada?: number
  cantidadSalida?: number
  cantidad?: number
  fechaIngreso: string | Date
  unidadMedidaId: string
  ubicacionId: number // ✅ NECESARIO para el formulario
  ubicacion: {
    rack: number
    fila: string
    posicion: string
  }
  usuarioReportado?: {
    nombre?: string
  }
  reportadoPorId: number
}